+++ branches/2.0/src/Partial/PartialView.php
@@ -39,9 +39,9 @@ class PartialView extends PlatesFactory implements PartialViewContract
     {
         if (in_array($name, $this->mixins)) {
             try {
-                $field = $this->engine->params('partial');
+                $driver = $this->engine->params('driver');
 
-                return $field->{$name}(...$args);
+                return $driver->{$name}(...$args);
             } catch (Exception $e) {
                 throw new BadMethodCallException(sprintf(
                     __CLASS__ . ' throws an exception during the method call [%s] with message : %s',