+++ branches/2.0/src/Wordpress/Metabox/Driver/Filefeed/Filefeed.php
@@ -2,6 +2,7 @@
 
 namespace tiFy\Wordpress\Metabox\Driver\Filefeed;
 
+use tiFy\Contracts\Metabox\MetaboxDriver as MetaboxDriverContract;
 use tiFy\Metabox\Driver\Filefeed\Filefeed as BaseFilefeed;
 
 class Filefeed extends BaseFilefeed
@@ -9,12 +10,14 @@ class Filefeed extends BaseFilefeed
     /**
      * @inheritDoc
      */
-    public function boot(): void
+    public function boot(): MetaboxDriverContract
     {
         parent::boot();
 
         add_action('admin_enqueue_scripts', function() {
             @wp_enqueue_media();
         });
+
+        return $this;
     }
 }