+++ branches/2.0/src/Wordpress/Metabox/Driver/Videofeed/Videofeed.php
@@ -2,6 +2,7 @@
 
 namespace tiFy\Wordpress\Metabox\Driver\Videofeed;
 
+use tiFy\Contracts\Metabox\MetaboxDriver as MetaboxDriverContract;
 use tiFy\Metabox\Driver\Videofeed\Videofeed as BaseVideofeed;
 
 class Videofeed extends BaseVideofeed
@@ -9,12 +10,14 @@ class Videofeed extends BaseVideofeed
     /**
      * @inheritDoc
      */
-    public function boot(): void
+    public function boot(): MetaboxDriverContract
     {
         parent::boot();
 
         add_action('admin_enqueue_scripts', function() {
             @wp_enqueue_media();
         });
+
+        return $this;
     }
 }